Реализована отправка писем о предстоящих записях
