+++ SoVet.Data/Handlers/Registration/SendMailCommandHandler.cs
@@ -41,7 +41,6 @@ public sealed class SendMailCommandHandler : IRequestHandler<SendMailCommand, Ba
             
             return await _emailService.SendEmailAsync(email, "Запись на прием", $"Здравствуйте,{client.Name}! " +
                 $"Хотим сообщить вам, что у вас присутсвует запись на {registration.StartTime}. Если вы не имеете связи с данным письмом, просим проигнорировать его. Ждем вас в нашей клинике!");
-            
         }
 
         return new BaseResponse { IsSuccess = true };
