+++ SoVet.Domain/Services/TimedHostedService.cs
@@ -1,0 +1,54 @@
+﻿using MediatR;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
+using SoVet.Domain.Commands.Registration;
+
+namespace SoVet.Domain.Services;
+
+public class TimedHostedService : IHostedService, IDisposable
+{
+    private int executionCount = 0;
+    private readonly ILogger<TimedHostedService> _logger;
+    private readonly IServiceProvider _services;
+    private Timer? _timer = null;
+
+    public TimedHostedService(ILogger<TimedHostedService> logger, IServiceProvider services)
+    {
+        _logger = logger;
+        _services = services;
+    }
+
+    private async void SendEmail(object? state)
+    {
+        _logger.LogInformation("Message sended");
+        using var scope = _services.CreateScope();
+        var mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
+        
+        var command = new SendMailCommand();
+        await mediator.Send(command);
+    }
+
+    public Task StartAsync(CancellationToken cancellationToken)
+    {
+        _logger.LogInformation("Timed Hosted Service running.");
+        _timer = new Timer(SendEmail, null, TimeSpan.Zero,
+            TimeSpan.FromSeconds(15));
+
+        return Task.CompletedTask;
+    }
+
+    public Task StopAsync(CancellationToken stoppingToken)
+    {
+        _logger.LogInformation("Timed Hosted Service is stopping.");
+
+        _timer?.Change(Timeout.Infinite, 0);
+
+        return Task.CompletedTask;
+    }
+
+    public void Dispose()
+    {
+        _timer?.Dispose();
+    }
+}
